fix typo in plugin name: "analisys" -> "analysis"
note that existing releases have this typo in the artifact name as well,
accordingly the links in `README.md` have been left unchanged.
(incidentally, this is how i noticed it: the artifact name looked
weird).

please note that that this PR is untested as the project cannot easily
be built locally due to the way the `build.gradle` is written with
release-checks as code in the middle of it (this should really be in a
task containing release checks). issue #2 has been raised for this.

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>