slightly simplify version handling in gradle build
so far there were two variables with extremely similar names:
* `opensearchVersion` defined in `gradle.properties` and containing the
  version of the plugin (and not of OpenSearch!)
* `opensearch_version` defined in `build.gradle` and containing the
  version of OpenSearch against which the plugin is built

now only one variable exists: `opensearchVersion`, defined in
`build.properties`. the version of the plugin is derived from this by
appending `.0` to it (e.g. for OpenSearch 2.2.0 the plugin version will
be 2.2.0.0). in the rare case that a new version of the plugin should be
published for an OpenSearch release for which a .0 version of the plugin
has already been published this will have to be bumped manually (and
reset for the next normal release).
the proper solution will be to eventually move to the standard build
from the plugin template (see issue #2).

Signed-off-by: Ralph Ursprung <Ralph.Ursprung@avaloq.com>